requirements.txt
numpy
scipy
trimesh
# Optional for training if you want GPU/ML: torch torchvision
PyYAML
